file location change
the file location for the osx-lock-on-remove-usb.sh file can not go into /opt/ there is not directory for macOS 10.13.6 (17G14033) and homebrew. I am aware of the unsupported outdated etc behavior language etc. This location change for cp the file allows for the program to work properly now. so for older MacOS users :)